h2_all_core/requirements.txt
africastalking==1.1.5
certifi==2019.3.9
chardet==3.0.4
contextlib2==0.5.5
dj-database-url==0.5.0
Django==2.2
django-stubs==0.12.1
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
idna==2.8
mypy==0.701
mypy-extensions==0.4.1
psycopg2-binary==2.8.2
pymongo==3.8.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
requests==2.21.0
schema==0.7.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
typing-extensions==3.7.2
urllib3==1.24.2
